move validation to a separate function, add tests
