+++ packages/babel-make-styles/src/index.ts
@@ -1,4 +1,4 @@
 export { configSchema } from './schema';
 
 export { plugin as default } from './plugin';
-export type { BabelPluginOptions } from './plugin';
+export type { BabelPluginOptions } from './types';
