+++ packages/babel-make-styles/src/types.ts
@@ -1,0 +1,12 @@
+import { TransformOptions } from '@babel/core';
+
+export type BabelPluginOptions = {
+  /** Defines set of modules and imports handled by a plugin. */
+  modules?: { moduleSource: string; importName: string }[];
+
+  /**
+   * If you need to specify custom Babel configuration, you can pass them here. These options will be used by the
+   * plugin when parsing and evaluating modules.
+   */
+  babelOptions?: Pick<TransformOptions, 'plugins' | 'presets'>;
+};
