+++ packages/babel-make-styles/src/validateOptions.ts
@@ -1,0 +1,17 @@
+import Ajv from 'ajv';
+
+import { configSchema } from './schema';
+import { BabelPluginOptions } from './types';
+
+const ajv = new Ajv();
+
+/**
+ * Validates options for plugin with a schema.
+ */
+export function validateOptions(pluginOptions: BabelPluginOptions): void {
+  const valid = ajv.validate(configSchema, pluginOptions);
+
+  if (!valid) {
+    throw new Error(`Validation failed for passed config: ${ajv.errorsText(ajv.errors)}`);
+  }
+}
